requirements.txt
streamlit
pdfplumber
python-docx
requests
